Refactor plumitas.py
