+++ plumitas/__init__.py
@@ -1,3 +1,5 @@
 from __future__ import absolute_import, division, print_function
 from .version import __version__  # noqa
-from .plumitas import *  # noqa
+from .core import *  # noqa
+from .analysis import *
+from .visualization import *
